refactor(figma): move component property types outside of react target
