+++ packages/figma/src/codegen/component-properties.ts
