+++ packages/figma/src/codegen/targets/react/component/transformers/avatar.ts
@@ -2,7 +2,10 @@ import { createElement, defineComponentTransformer } from "@/codegen/core";
 import * as metadata from "@/entities/data/__generated__/component-sets";
 import { findAllInstances } from "@/utils/figma-node";
 import type { SeedComponentTransformerDeps } from "../deps.interface";
-import type { AvatarProperties, IdentityPlaceholderProperties } from "../properties.type";
+import type {
+  AvatarProperties,
+  IdentityPlaceholderProperties,
+} from "@/codegen/component-properties";
 import { createIdentityPlaceholderTransformer } from "./identity-placeholder";
 
 export const createAvatarTransformer = (ctx: SeedComponentTransformerDeps) => {
