+++ packages/figma/src/codegen/targets/react/component/transformers/badge.ts
@@ -2,7 +2,7 @@ import { createElement, defineComponentTransformer } from "@/codegen/core";
 import * as metadata from "@/entities/data/__generated__/component-sets";
 import { camelCase } from "change-case";
 import type { SeedComponentTransformerDeps } from "../deps.interface";
-import type { BadgeProperties } from "../properties.type";
+import type { BadgeProperties } from "@/codegen/component-properties";
 import { handleSizeProp } from "../size";
 
 export const createBadgeTransformer = (_ctx: SeedComponentTransformerDeps) =>
