+++ packages/figma/src/codegen/targets/react/component/transformers/chip-tabs.ts
@@ -3,7 +3,7 @@ import * as metadata from "@/entities/data/__generated__/component-sets";
 import { findAllInstances } from "@/utils/figma-node";
 import { camelCase } from "change-case";
 import type { SeedComponentTransformerDeps } from "../deps.interface";
-import type { ChipTabsItemProperties, ChipTabsProperties } from "../properties.type";
+import type { ChipTabsItemProperties, ChipTabsProperties } from "@/codegen/component-properties";
 
 const CHIP_TABS_ITEM_KEY = "fa80168b02051fbb0ba032238bd76d840dbe2e15";
 const createChipTabsItemTransformer = (_ctx: SeedComponentTransformerDeps) =>
