+++ packages/figma/src/codegen/targets/react/component/transformers/extended-fab.ts
@@ -2,7 +2,7 @@ import { createElement, defineComponentTransformer } from "@/codegen/core";
 import * as metadata from "@/entities/data/__generated__/component-sets";
 import { camelCase } from "change-case";
 import type { SeedComponentTransformerDeps } from "../deps.interface";
-import type { ExtendedFabProperties } from "../properties.type";
+import type { ExtendedFabProperties } from "@/codegen/component-properties";
 import { handleSizeProp } from "../size";
 
 export const createExtendedFabTransformer = (ctx: SeedComponentTransformerDeps) =>
