+++ packages/figma/src/codegen/targets/react/component/transformers/identity-placeholder.ts
@@ -2,7 +2,7 @@ import { defineComponentTransformer } from "@/codegen/core";
 import { camelCase } from "change-case";
 import * as metadata from "@/entities/data/__generated__/component-sets";
 import { createElement } from "@/codegen/core";
-import type { IdentityPlaceholderProperties } from "../properties.type";
+import type { IdentityPlaceholderProperties } from "@/codegen/component-properties";
 import type { SeedComponentTransformerDeps } from "../deps.interface";
 
 export const createIdentityPlaceholderTransformer = (_ctx: SeedComponentTransformerDeps) =>
